backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
supabase==2.0.2
openai==1.3.5
pypdf==3.17.1
numpy==1.24.3
python-dotenv==1.0.0
cryptography>=3.1
playwright>=1.40.0
